Scrap scala project in favour of flippy-standalone
Made a more robust version of the actual service at flippy-standalone,
but this project is still useful for the docker compose and nginx work.

Pull in flippy-standalone instead of building a service, and shuffle the
nginx config a bit to be more sensibly placed (even when there was a
maven project here, it didn't really belong in src/main/resources)

It'll still pull the admin from ../flippy, but this is only a test
project still, so that's not a big deal. In production it'll be up to
the deployer to pull in the admin static files and host them where they
see fit, and this project will serve as a guide as to how that may be
done.